+++ src/Components/Layout/index.js
@@ -1,14 +1,14 @@
-import Header from '../Header/index';
-import Navbar from '../Navbar';
-import Footer from '../Footer/index';
+import styles from './layout.module.css';
+import Header from 'Components/Shared/Header/index';
+import Navbar from 'Components/Shared/Navbar';
+import Footer from 'Components/Shared/Footer/index';
+import Home from 'Components/Shared/Home/index';
 import Admins from '../Admins/index';
 import AdminsForm from '../Admins/Form';
-import SuperAdmins from '../SuperAdmins/index';
-import SuperAdminsForm from '../SuperAdmins/Form';
-import Home from '../Home/index';
-import styles from './layout.module.css';
 import Employees from '../Employees/index';
 import EmployeesForm from '../Employees/Form';
+import SuperAdmins from '../SuperAdmins/index';
+import SuperAdminsForm from '../SuperAdmins/Form';
 import Projects from '../Projects';
 import ProjectsForm from '../Projects/Form';
 import TimeSheets from '../TimeSheets';
